+++ Monopoly/src/Event.java
@@ -1,8 +1,10 @@
 public abstract class Event extends Tile
 	{
+		private static int price;
+
 		public Event(String name)
 			{
-				super(name);
+				super(name, price);
 
 			}
