+++ Monopoly/src/Tile.java
@@ -2,10 +2,12 @@
 public abstract class Tile
 	{
 		protected String name;
+		protected int price;
 
-		public Tile(String n)
+		public Tile(String n, int p)
 			{
 				name = n;
+				price = p;
 			}
 
 		public void landOn()
